use signed horizontal and vertical offsets in clap

+++ src/isomedia/box_code_base.c
@@ -8313,9 +8313,9 @@ GF_Err clap_box_read(GF_Box *s, GF_BitStream *bs)
 	ptr->cleanApertureWidthD = gf_bs_read_u32(bs);
 	ptr->cleanApertureHeightN = gf_bs_read_u32(bs);
 	ptr->cleanApertureHeightD = gf_bs_read_u32(bs);
-	ptr->horizOffN = gf_bs_read_u32(bs);
+	ptr->horizOffN = (s32) gf_bs_read_u32(bs);
 	ptr->horizOffD = gf_bs_read_u32(bs);
-	ptr->vertOffN = gf_bs_read_u32(bs);
+	ptr->vertOffN = (s32) gf_bs_read_u32(bs);
 	ptr->vertOffD = gf_bs_read_u32(bs);
 	return GF_OK;
 }
@@ -8332,9 +8332,9 @@ GF_Err clap_box_write(GF_Box *s, GF_BitStream *bs)
 	gf_bs_write_u32(bs, ptr->cleanApertureWidthD);
 	gf_bs_write_u32(bs, ptr->cleanApertureHeightN);
 	gf_bs_write_u32(bs, ptr->cleanApertureHeightD);
-	gf_bs_write_u32(bs, ptr->horizOffN);
+	gf_bs_write_u32(bs, (u32) ptr->horizOffN);
 	gf_bs_write_u32(bs, ptr->horizOffD);
-	gf_bs_write_u32(bs, ptr->vertOffN);
+	gf_bs_write_u32(bs, (u32) ptr->vertOffN);
 	gf_bs_write_u32(bs, ptr->vertOffD);
 	return GF_OK;
 }
